Moves the logic for making contact cards into the Rolodex view
